% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ace27.R
\name{assign_ace27}
\alias{assign_ace27}
\title{Assign ACE-27}
\usage{
assign_ace27(
  ace_mi,
  ace_cad,
  ace_chf,
  ace_arr,
  ace_htn,
  ace_vd,
  ace_pad,
  ace_res,
  ace_hep,
  ace_sto,
  ace_pan,
  ace_rd,
  ace_dm,
  ace_str,
  ace_dem,
  ace_par,
  ace_neu,
  ace_psy,
  ace_rhe,
  ace_aid,
  ace_st,
  ace_lm,
  ace_lym,
  ace_alc,
  ace_id,
  ace_obe,
  separate_nos = FALSE
)
}
\arguments{
\item{ace_mi}{Myocardial Infarction}

\item{ace_cad}{Angina / Coronary Heart Disease}

\item{ace_chf}{Congestive Heart Failure}

\item{ace_arr}{Arrhythmias}

\item{ace_htn}{Hypertension}

\item{ace_vd}{Venous Disease}

\item{ace_pad}{Peripheral Arterial Disease}

\item{ace_res}{Respiratory System}

\item{ace_hep}{Hepatic}

\item{ace_sto}{Stomach / Intestine}

\item{ace_pan}{Pancreas}

\item{ace_rd}{End-stage renal disease}

\item{ace_dm}{Diabetes Mellitus}

\item{ace_str}{Stroke}

\item{ace_dem}{Dementia}

\item{ace_par}{Paralysis}

\item{ace_neu}{Neuromuscular}

\item{ace_psy}{Psychiatric}

\item{ace_rhe}{Rheumatologic}

\item{ace_aid}{AIDS}

\item{ace_st}{Solid Tumor including melanoma}

\item{ace_lm}{Leukemia and Myeloma}

\item{ace_lym}{Lymphoma}

\item{ace_alc}{Alcohol}

\item{ace_id}{Illicit Drugs}

\item{ace_obe}{Obesity}

\item{separate_nos}{logical; if \code{TRUE}, cases that only have "NOS" ailments
(excluding Obesity, NOS) are separated into their own category.}
}
\value{
A factor with levels "None", ">=Mild", ">=Moderate", "Severe" (and
"NOS" if \code{separate_nos = TRUE}) indicating the overall ACE-27 score.
}
\description{
Assign Adult Comorbidity Evaluation-27 index based on 26 medical
comorbidities.
}
\details{
Each comorbidity is scored on a scale of Grades 0-3, where 0 = None, 1 =
Mild, 2 = Moderate, and 3 = Severe. The ACE-27 overall score is assigned
according to the highest ranked individual comorbidity recorded. However, if
there are two or more Grade 2 comorbidities occuring in different organ
systems, then the overall score becomes Grade 3.

When there is insufficient information regarding a comorbidity, then it is
described as an "NOS" (not otherwise specified) case. There are two ways to
handle these two cases. First, we can set \code{separate_nos = TRUE} to create a
separate category of cases that \emph{only} have NOS ailments, with "Obesity, NOS"
being the exception as it is only assigned to Grade 2. Or, we can set
\code{separate_nos = FALSE} and have the NOS cases be categorized with the lowest
risk group based on information we have on all the ACE variables. Thus,
we relabel "Mild" and "Moderate" to ">=Mild" and ">=Moderate" respectively
to reflect the risk uncertainty.

Two other scoring assumptions are made. First, if there is insufficient
information regarding a comorbidity then it is described as an "NOS" (not
otherwise specified) case, and scored as Grade 1. To account for the
possibility an NOS case is actually higher than Grade 1, all Grade 1/2 are
relabelled with ">=". Alternatively, one can also specify \code{separate_nos} as
\code{TRUE} to create a separate category of cases that \emph{only} have NOS ailments,
with "Obesity, NOS" being the exception as it is only assigned to Grade 2.

Any case with multiple symptoms from different Grades is scored
into the higher Grade. For example, someone with symptoms from both Grade 1
and Grade 2 is categorized as Grade 2.
\subsection{ACE comorbidities by organ systems}{
\itemize{
\item Cardiovascular System
\itemize{
\item Myocardial Infarction
\item Angina / Coronary Heart Disease
\item Congestive Heart Failure
\item Arrhythmias
\item Hypertension
\item Venous Disease
\item Peripheral Arterial Disease
}
\item Respiratory System
\item Gastrointestinal System
\itemize{
\item Hepatic
\item Stomach / Intestine
\item Pancreas
}
\item Renal System
\itemize{
\item End-stage renal disease
}
\item Endocrine System
\itemize{
\item Diabetes Mellitus
}
\item Neurological System
\itemize{
\item Stroke
\item Dementia
\item Paralysis
\item Neuromuscular
}
\item Psychiatric
\item Rheumatologic
\item Immunological System
\itemize{
\item AIDS
}
\item Malignancy
\itemize{
\item Solid Tumor including melanoma
\item Leukemia and Myeloma
\item Lymphoma
}
\item Substance Abuse
\itemize{
\item Alcohol
\item Illicit Drugs
}
\item Body Weight
\itemize{
\item Obesity
}
}
}
}
\references{
Binder PS, Peipert JF, Kallogjeri D, et al. Adult Comorbidity
Evaluation 27 score as a predictor of survival in endometrial cancer
patients. Am J Obstet Gynecol. 2016;215(6):766.e1-766.e9.
doi:10.1016/j.ajog.2016.07.035

https://siog.org/files/public/ace27.pdf
}
\author{
Derek Chiu
}
