% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ace27.R
\name{assign_ace_vars}
\alias{assign_ace_vars}
\title{Assign ACE variables into numeric scores}
\usage{
assign_ace_vars(
  ace_mi,
  ace_cad,
  ace_chf,
  ace_arr,
  ace_htn,
  ace_vd,
  ace_pad,
  ace_res,
  ace_hep,
  ace_sto,
  ace_pan,
  ace_rd,
  ace_dm,
  ace_str,
  ace_dem,
  ace_par,
  ace_neu,
  ace_psy,
  ace_rhe,
  ace_aid,
  ace_st,
  ace_lm,
  ace_lym,
  ace_alc,
  ace_id,
  ace_obe
)
}
\arguments{
\item{ace_mi}{Myocardial Infarction}

\item{ace_cad}{Angina / Coronary Heart Disease}

\item{ace_chf}{Congestive Heart Failure}

\item{ace_arr}{Arrhythmias}

\item{ace_htn}{Hypertension}

\item{ace_vd}{Venous Disease}

\item{ace_pad}{Peripheral Arterial Disease}

\item{ace_res}{Respiratory System}

\item{ace_hep}{Hepatic}

\item{ace_sto}{Stomach / Intestine}

\item{ace_pan}{Pancreas}

\item{ace_rd}{End-stage renal disease}

\item{ace_dm}{Diabetes Mellitus}

\item{ace_str}{Stroke}

\item{ace_dem}{Dementia}

\item{ace_par}{Paralysis}

\item{ace_neu}{Neuromuscular}

\item{ace_psy}{Psychiatric}

\item{ace_rhe}{Rheumatologic}

\item{ace_aid}{AIDS}

\item{ace_st}{Solid Tumor including melanoma}

\item{ace_lm}{Leukemia and Myeloma}

\item{ace_lym}{Lymphoma}

\item{ace_alc}{Alcohol}

\item{ace_id}{Illicit Drugs}

\item{ace_obe}{Obesity}
}
\description{
Each ACE variable is assigned into numeric scores ranging from 0 to 3.
}
