% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_esmo2016.R
\name{assign_esmo2016}
\alias{assign_esmo2016}
\title{Assign ESMO 2016}
\usage{
assign_esmo2016(stage, grade, hist, myo, lvi)
}
\arguments{
\item{stage}{FIGO stage}

\item{grade}{tumour grade}

\item{hist}{histotype}

\item{myo}{myometrial invasion}

\item{lvi}{lymphovascular space invasion}
}
\value{
ESMO 2016 assigned into "low", "intermediate", "high-intermediate",
"high", "advanced", or "metastatic". Unassignable cases are \code{NA}.
}
\description{
Assign ESMO 2016 based on stage, grade, histotype, myometrial invasion, and
LVSI.
}
\details{
ESMO 2013 is assigned using stage, grade, histotype, myometrial invasion, and
LVSI into low, intermediate, high-intermediate, high, advanced, and
metastatic risk based on the following criteria:
\itemize{
\item low:
\itemize{
\item stage I, grade 1/2, endometrioid, <50\% myometrial invasion, LVSI negative
}
\item intermediate:
\itemize{
\item stage I, grade 1/2, endometrioid, >=50\% myometrial invasion, LVSI negative
}
\item high-intermediate:
\itemize{
\item stage I, grade 3, endometrioid, <50\% myometrial invasion, regardless of LVSI status
\item stage I, grade 1/2, endometrioid, LVSI unequivocally positive, regardless of depth of invasion
}
\item high:
\itemize{
\item stage I, grade 3, endometrioid, >=50\% myometrial invasion, regardless of LVSI status
\item stage II or higher
\item stage III, endometrioid, no residual disease
\item non-endometrioid (serous or clear-cell or undifferentiated carcinoma, or carcinosarcoma)
}
\item advanced:
\itemize{
\item stage III residual disease
\item stage IVA
}
\item metastatic:
\itemize{
\item stage IVB
}
}
}
\note{
Assignment starts from the metastatic group first as the criteria are
not mutually exclusive. Residual disease is not available and so are not
used in criteria for high-intermediate and high.
}
\references{
Colombo et al. Annals of Oncology 2016
}
\author{
Derek Chiu, Samuel Leung
}
