% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mmr2.R
\name{assign_mmr2}
\alias{assign_mmr2}
\title{Assign MMR status by 2 IHC markers}
\usage{
assign_mmr2(msh6, pms2)
}
\arguments{
\item{msh6}{MSH6 marker}

\item{pms2}{PMS2 marker}
}
\value{
MMR status, either "intact" or "deficient"
}
\description{
Assign MMR status by 2 IHC markers, MSH6 and PMS2.
}
\details{
MMR status, intact or deficient, is determined by IHC markers MSH6 and PMS2
consolidated from TMA scoring. If both markers are intact, then MMR status is
intact. Otherwise, MMR status is deficient (including subclonal and unknown
markers).
}
\author{
Samuel Leung, Derek Chiu
}
