% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mmr4.R
\name{assign_mmr4}
\alias{assign_mmr4}
\title{Assign MMR status by 4 IHC markers}
\usage{
assign_mmr4(msh6, pms2, mlh1, msh2)
}
\arguments{
\item{msh6}{MSH6 marker}

\item{pms2}{PMS2 marker}

\item{mlh1}{MLH1 marker}

\item{msh2}{MSH2 marker}
}
\value{
MMR status, either "intact" or "deficient"
}
\description{
Assign MMR status by 4 IHC markers, MSH6, PMS2, MLH1, and MSH2.
}
\details{
MMR status, intact or deficient, is determined by IHC markers MSH6, PMS2,
MLH1, and MSH2 consolidated from TMA scoring. If all markers are intact, then
MMR status is intact. Otherwise, MMR status is deficient (including subclonal
and unknown markers).
}
\examples{
mmr_ihc_4 <- with(emdb, assign_mmr4(msh6, pms2, mlh1, msh2))
table(mmr_ihc_4)
}
\author{
Samuel Leung, Derek Chiu
}
