% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_promise2015.R
\name{assign_promise2015}
\alias{assign_promise2015}
\title{Assign ProMisE 2015}
\usage{
assign_promise2015(mmr, pole, p53)
}
\arguments{
\item{mmr}{MMR status determined by 2 IHCs: MSH6 and PMS2}

\item{pole}{POLE mutation status}

\item{p53}{p53 IHC}
}
\value{
ProMisE 2015 assigned into "MMRd", "POLEmut", "p53abn", or "NSMP/p53wt"
}
\description{
Assign ProMisE 2015 based on MMR, POLE, and p53.
}
\details{
ProMisE 2015 is assigned in the following order:
\enumerate{
\item MMR deficient cases if either of the 2 IHCs MSH6 or PMS2 are absent
\item POLE all mutations (pathogenic and non-pathogenic) among MMR intact cases
\item p53 IHC wild type or mutated among POLE wild type cases
}
}
\author{
Derek Chiu, Samuel Leung
}
