% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_promise2019.R
\name{assign_promise2019}
\alias{assign_promise2019}
\title{Assign ProMisE 2019}
\usage{
assign_promise2019(mmr_ihc_2, pole_mut, p53)
}
\arguments{
\item{mmr_ihc_2}{MMR status determined by 2 IHCs: MSH6 and PMS2}

\item{pole_mut}{POLE mutation status}

\item{p53}{p53 IHC}
}
\value{
ProMisE 2019 assigned into "POLEmut", "MMRd", "p53abn", or "NSMP/p53wt"
}
\description{
Assign ProMisE 2015 based on POLE, MMR, and p53.
}
\details{
ProMisE 2015 is assigned in the following order:
\enumerate{
\item POLE pathogenic mutations
\item MMR deficient cases if either of the 2 IHCs MSH6 or PMS2 are absent,
among POLE non-pathogenic mutations or wild type cases
\item p53 IHC wild type or mutated among MMR intact cases
}
}
\examples{
promise2019 <- with(emdb, assign_promise2019(mmr_ihc_2, pole_mut, p53))
table(promise2019)
}
\author{
Derek Chiu, Samuel Leung
}
