% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_promisepoor.R
\name{assign_promisepoor}
\alias{assign_promisepoor}
\title{Assign Poor-man's ProMisE}
\usage{
assign_promisepoor(mmr2, mmr4, pole, p53, grade, hist_gr)
}
\arguments{
\item{mmr2}{MMR status determined by 2 IHCs: MSH6 and PMS2}

\item{mmr4}{MMR status determined by 4 IHCs: MSH6, PMS2, MLH1, and MSH2}

\item{pole}{POLE mutation status}

\item{p53}{p53 IHC}

\item{grade}{tumour grade: 1, 2, 3}

\item{hist_gr}{histological subtype group: endometrioid or non-endometrioid}
}
\value{
Poor-man's ProMisE assigned into "Low-Risk", "POLEmut", "MMRd",
"p53abn", or "NSMP/p53wt"
}
\description{
Assign Poor-man's ProMisE based on POLE, MMR, p53, and histopathological
features.
}
\details{
Poor-man's ProMisE first assigns a low risk group based on histopathological
features grade, histotype, MMR status determined by 2 and 4 IHC markers, and
p53. The rest of the high risk cases are assigned as in ProMisE 2019.
}
\author{
Derek Chiu, Samuel Leung
}
