% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check input parameters}
\usage{
check_input(param, codelist)
}
\arguments{
\item{param}{input parameter}

\item{codelist}{vector of accepted values for the input parameter}
}
\value{
if input parameter contains only accepted values, return \code{TRUE};
otherwise, an error message is displayed indicating the invalid values in
\code{param}
}
\description{
Check input parameters belong to accepted code list for data validation.
}
\details{
Each clinical variable has a specific list of accepted values. If an invalid
value is provided, e.g. "IE" in stage, then an error should be thrown. The
purpose is to prevent situations where classification algorithms
such as ESMO may assign "unassignable" for these cases, and instead flag
the user of invalid input parameters.

The list of accepted values for each clinical variable are stored as
constants in the package.
}
\examples{
STAGE_STD <- c("I", "IA", "IB", "IC", "II", "IIA", "IIB", "III", "IIIA",
"IIIB", "IIIC", "IIIC1", "IIIC2", "IV", "IVA", "IVB")
check_input(emdb$stage_full, STAGE_STD)
}
\author{
Derek Chiu
}
