% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_esmo2013.R
\name{assign_esmo2013}
\alias{assign_esmo2013}
\title{Assign ESMO 2013}
\usage{
assign_esmo2013(stage, grade, hist_gr)
}
\arguments{
\item{stage}{FIGO stage: I, II, III, IV with substages}

\item{grade}{tumour grade: 1, 2, 3}

\item{hist_gr}{histological subtype group: endometrioid or non-endometrioid}
}
\value{
ESMO 2013 assigned into "low", "intermediate", or "high".
Unassignable cases are \code{NA}.
}
\description{
Assign ESMO 2013 based on stage, grade, and histological subtype group
}
\details{
ESMO 2013 is assigned using stage, grade, and histological subtype group into
low, intermediate, and high risk based on the following criteria:
\itemize{
\item low:
\itemize{
\item stage IA, grade 1/2, with endometrioid type
}
\item intermediate:
\itemize{
\item stage IA, grade 3, with endometrioid type
\item stage IB, grade 1/2, with endometrioid type
}
\item high:
\itemize{
\item all stages with non-endometrioid type
\item stage IB, grade 3, with endometrioid type
\item stage II or higher
}
}
}
\note{
Assignment starts from the high group first as the criteria are not
mutually exclusive.
}
\examples{
esmo2013 <- with(emdb, assign_esmo2013(stage_full, grade_rev, hist_rev_gr))
table(esmo2013)
}
\references{
Colombo et al. Ann Oncol 2013
}
\author{
Derek Chiu, Samuel Leung
}
