% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_esmo2020.R
\name{assign_esmo2020}
\alias{assign_esmo2020}
\title{Assign ESMO 2020}
\usage{
assign_esmo2020(stage, grade, hist_gr, hist, myo, lvi, eclass = NULL)
}
\arguments{
\item{stage}{FIGO stage: I, II, III, IV with substages}

\item{grade}{tumour grade: 1, 2, 3}

\item{hist_gr}{histological subtype group: endometrioid or non-endometrioid}

\item{hist}{histological subtype: detailed histotype}

\item{myo}{myometrial invasion: none, 1-50\%, >50\%}

\item{lvi}{lymphovascular space invasion: negative, positive, focal, extensive}

\item{eclass}{molecular classification: "MMRd", "POLEmut", "p53abn", or
"NSMP/p53wt"}
}
\value{
ESMO 202020 assigned into "low", "intermediate", "high-intermediate",
"high", "advanced", or "metastatic". Unassignable cases are \code{NA}.
}
\description{
Assign ESMO 2020 based on stage, grade, histological subtype group,
histological subtype, myometrial invasion, and LVSI. Molecular classification
can be used if available.
}
\details{
ESMO 2020 is assigned using stage, grade, histological subtype group,
histological subtype, myometrial invasion, and LVSI into low, intermediate,
high-intermediate, high, advanced, and metastatic risk based on the following
criteria with and without molecular classification:

\strong{With molecular classification}
\itemize{
\item low:
\itemize{
\item stage I-II, POLEmut
\item stage IA, grade 1/2, endometrioid, <50\% myometrial invasion, LVSI negative or focal, MMRd/NSMP
}
\item intermediate:
\itemize{
\item stage IB, grade 1/2, endometrioid, <50\% myometrial invasion, LVSI negative or focal, MMRd/NSMP
\item stage IA, grade 3, endometrioid, LVSI negative or focal, MMRd/NSMP
\item stage IA, non-endometrioid or mixed subtype (serous, carcinosarcoma (MMMT), or undifferentiated), no myometrial invasion, p53abn
}
\item high-intermediate:
\itemize{
\item stage IA/IB, endometrioid, LVSI positive or extensive, MMRd/NSMP
\item stage IB, grade 3, endometrioid, MMRd/NSMP
\item stage II/IIA, endometrioid, MMRd/NSMP
}
\item high:
\itemize{
\item stage III-IVA, endometrioid, MMRd/NSMP
\item stage I-IVA, endometrioid, >0\% myometrial invasion, p53abn
\item stage IA, non-endometrioid or mixed subtype (serous, carcinosarcoma (MMMT), or undifferentiated), >0\% myometrial invasion, MMRd/NSMP
\item stage IB-IVA, non-endometrioid or mixed subtype (serous, carcinosarcoma (MMMT), or undifferentiated), MMRd/NSMP
}
\item advanced:
\itemize{
\item stage III-IVA, any molecular type
}
\item metastatic:
\itemize{
\item stage IVB, any molecular type
}
}

\strong{Without molecular classification}
\itemize{
\item low:
\itemize{
\item stage IA, grade 1/2, endometrioid, <50\% myometrial invasion, LVSI negative or focal
}
\item intermediate:
\itemize{
\item stage IB, grade 1/2, endometrioid, <50\% myometrial invasion, LVSI negative or focal
\item stage IA, grade 3, endometrioid, LVSI negative or focal
\item stage IA, non-endometrioid or mixed subtype (serous, carcinosarcoma (MMMT), or undifferentiated), no myometrial invasion
}
\item high-intermediate:
\itemize{
\item stage IA/IB, endometrioid, LVSI positive or extensive
\item stage IB, grade 3, endometrioid
\item stage II/IIA, endometrioid
}
\item high:
\itemize{
\item stage III-IVA, endometrioid
\item stage IA, non-endometrioid or mixed subtype (serous, carcinosarcoma (MMMT), or undifferentiated), >0\% myometrial invasion
\item stage IB-IVA, non-endometrioid or mixed subtype (serous, carcinosarcoma (MMMT), or undifferentiated)
}
\item advanced:
\itemize{
\item stage III-IVA
}
\item metastatic:
\itemize{
\item stage IVB
}
}
}
\note{
Assignment starts from the low group first.
}
\references{
Concin N, Matias-Guiu X, Vergote I, et al ESGO/ESTRO/ESP
guidelines for the management of patients with endometrial
carcinoma. International Journal of Gynecologic Cancer Published Online
First: 18 December 2020. doi: 10.1136/ijgc-2020-002230
}
\author{
Derek Chiu
}
