% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_esmo2020.R
\name{assign_esmo2020}
\alias{assign_esmo2020}
\title{Assign ESMO 2020}
\usage{
assign_esmo2020(
  stage_full,
  grade,
  hist_gr,
  myo,
  lvi,
  eclass = NULL,
  residual = NULL
)
}
\arguments{
\item{stage_full}{FIGO stage: I, II, III, IV with substages}

\item{grade}{tumour grade: 1, 2, 3}

\item{hist_gr}{histological subtype group: endometrioid or non-endometrioid}

\item{myo}{myometrial invasion: none, 1-50\%, >50\%}

\item{lvi}{lymphovascular space invasion: negative, positive, focal, extensive}

\item{eclass}{molecular classification: "MMRd", "POLEmut", "p53abn", or
"NSMP/p53wt"}

\item{residual}{residual disease: "<1cm", ">=1cm", "microscopic", or "no
residual". Use if available. Microscopic residual is regarded as no
residual disease.}
}
\value{
ESMO 2020 assigned into "low", "intermediate", "high-intermediate",
"high", "advanced", or "metastatic". Unassignable cases are \code{NA}.
}
\description{
Assign ESMO 2020 based on stage, grade, histological subtype group,
myometrial invasion, and LVSI. Molecular classification and residual disease
can be used if available.
}
\details{
ESMO 2020 is assigned using stage, grade, histological subtype group,
myometrial invasion, and LVSI into low, intermediate, high-intermediate,
high, advanced, and metastatic risk based on the following criteria with and
without molecular classification/residual disease:

\strong{With molecular classification}
\itemize{
\item low:
\itemize{
\item stage I-II, POLEmut, (no residual disease if known)
\itemize{
\item missing residual disease is regarded as no residual disease
}
\item stage IA, grade 1/2, endometrioid, LVSI negative or focal, MMRd/NSMP
}
\item intermediate:
\itemize{
\item stage IB, grade 1/2, endometrioid, LVSI negative or focal, MMRd/NSMP
\item stage IA, grade 3, endometrioid, LVSI negative or focal, MMRd/NSMP
\item stage IA, no myometrial invasion, p53abn
}
\item high-intermediate:
\itemize{
\item stage IA/IB, endometrioid, LVSI positive or extensive, MMRd/NSMP
\item stage IB, grade 3, endometrioid, MMRd/NSMP
\item stage II/IIA, endometrioid, MMRd/NSMP
}
\item high:
\itemize{
\item stage III-IVA, endometrioid, MMRd/NSMP, (no residual disease if known)
\item stage I-IVA, >0\% myometrial invasion, p53abn, (no residual disease if known)
\itemize{
\item for stage I-II, missing residual disease is regarded as no residual disease
}
\item stage IA, non-endometrioid or mixed subtype, >0\% myometrial invasion, MMRd/NSMP, (no residual disease if known)
\itemize{
\item missing residual disease is regarded as no residual disease
}
\item stage IB-IVA, non-endometrioid or mixed subtype, MMRd/NSMP, (no residual disease if known)
\itemize{
\item for stage IB-II, missing residual disease is regarded as no residual disease
}
}
\item advanced:
\itemize{
\item stage III-IVA, any molecular type, (with residual disease if known)
}
\item metastatic:
\itemize{
\item stage IVB, any molecular type
}
}

\strong{Without molecular classification}
\itemize{
\item low:
\itemize{
\item stage IA, grade 1/2, endometrioid, LVSI negative or focal
}
\item intermediate:
\itemize{
\item stage IB, grade 1/2, endometrioid, LVSI negative or focal
\item stage IA, grade 3, endometrioid, LVSI negative or focal
\item stage IA, no myometrial invasion
}
\item high-intermediate:
\itemize{
\item stage IA/IB, endometrioid, LVSI positive or extensive
\item stage IB, grade 3, endometrioid
\item stage II/IIA, endometrioid
}
\item high:
\itemize{
\item stage III-IVA, endometrioid, (no residual disease if known)
\item stage IA, non-endometrioid or mixed subtype, >0\% myometrial invasion, (no residual disease if known)
\itemize{
\item missing residual disease is regarded as no residual disease
}
\item stage IB-IVA, non-endometrioid or mixed subtype, (no residual disease if known)
\itemize{
\item for stage IB-II, missing residual disease is regarded as no residual disease
}
}
\item advanced:
\itemize{
\item stage III-IVA, (with residual disease if known)
}
\item metastatic:
\itemize{
\item stage IVB
}
}
}
\note{
Assignment starts from the low group first.
}
\examples{
# without molecular classification
esmo2020_wo_eclass <- with(emdb, assign_esmo2020(stage_full, grade_rev,
hist_rev_gr, myo, lvi))
table(esmo2020_wo_eclass)

# with molecular classification
eclass <- with(emdb, assign_promise2015(mmr_ihc_2, pole_mut, p53))
esmo2020_w_eclass <- with(emdb, assign_esmo2020(stage_full, grade_rev,
hist_rev_gr, myo, lvi, eclass))
table(esmo2020_w_eclass)

# with molecular classification and residual disease
esmo2020_w_eclass_res <- with(emdb, assign_esmo2020(stage_full, grade_rev,
hist_rev_gr, myo, lvi, eclass, residual))
table(esmo2020_w_eclass_res)
}
\references{
Concin N, Matias-Guiu X, Vergote I, et al ESGO/ESTRO/ESP
guidelines for the management of patients with endometrial
carcinoma. International Journal of Gynecologic Cancer Published Online
First: 18 December 2020. doi: 10.1136/ijgc-2020-002230
}
\author{
Derek Chiu
}
