% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emdb}
\alias{emdb}
\title{Simulated endometrial cancer Database}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 800 rows and 16 columns.
}
\usage{
emdb
}
\description{
Simulated endometrial cancer database used for testing functions.
}
\details{
Each clinical variable has a distribution simulated from the proportions seen
in the Vancouver endometrial cancer database.
}
\keyword{datasets}
