% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoTools-package.R
\docType{package}
\name{EndoTools-package}
\alias{EndoTools}
\alias{EndoTools-package}
\title{EndoTools: Endometrial Cancer Helper Tools}
\description{
Helper tools such as calculation of ProMisE classifications and ESMO risk groups for endometrial cancers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://talhouklab.github.io/EndoTools}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

Authors:
\itemize{
  \item Samuel Leung \email{Samuel.Leung@vch.ca}
}

}
\keyword{internal}
